Add files via upload
Real-Time Data Analysis and Prediction in Borivali, Kandivali, and Malad

Description:
This project focuses on real-time data analysis and prediction in the Borivali, Kandivali, and Malad regions. The project involves making necessary changes to the real-time data collected from these areas and performing data cleaning and preprocessing to ensure the data is in a usable and correct format.

Key Features:

Real-time data collection and processing
Data cleaning and preprocessing
Implementation of the Stacking ensemble technique
Use of SVR, Linear Regression, Ridge, Lasso, Random Forest, GBM, ENet, and Linear SVR models
Open for feedback and improvements
Purpose:
The project aims to provide a clear understanding of real-time data analysis and prediction techniques. It serves as a practical example for those looking to learn about data preprocessing and ensemble techniques. Feedback and contributions to improve the project are welcome.

Usage: